refactor edges
