+++ assets/src/components/component/tree/getTreeNodesAndEdges.tsx
@@ -2,7 +2,7 @@ import { ComponentTreeFragment, MetadataFragment } from 'generated/graphql'
 import { ConditionalKeys } from 'type-fest'
 import { isNonNullable } from 'utils/isNonNullable'
 
-import { STEP_EDGE_NAME } from '../../utils/ReactFlowEdges'
+import { DIRECTED_EDGE_NAME } from '../../utils/reactflow/edges'
 
 export type TreeNodeMeta = ReturnType<typeof flattenMetadata>[number]
 export type HasMetadata = { metadata?: MetadataFragment }
@@ -24,7 +24,7 @@ export const C_TYPES = [
 ] as const satisfies ComponentKindsKey[]
 
 export const baseEdgeProps = {
-  type: STEP_EDGE_NAME,
+  type: DIRECTED_EDGE_NAME,
   updatable: false,
 }
